requirements.txt

langchain
langchain-community
langchain-groq
groq
faiss-cpu
sentence-transformers
pypdf
python-dotenv
google-colab
streamlit
requests
-e .
